+++ ui/src/views/index.js
@@ -3,7 +3,6 @@ export { default as ClaudeSettingsView } from './ClaudeSettingsView';
 export { default as GeminiSettingsView } from './GeminiSettingsView';
 export { default as CreateMcpView } from './CreateMcpView';
 export { default as EnvView } from './EnvView';
-export { default as TemplatesView } from './TemplatesView';
 export { default as SubprojectsView } from './SubprojectsView';
 export { default as RegistryView } from './RegistryView';
 export { default as MemoryView } from './MemoryView';